Refactor on pixel class: - RxPixel has the responsability of sending, enqueueing, and subscribe to hide rxjava detail from the caller - PixelSender has the logic to send pixels to api, enqueue in database, and mapping values - PixelSender has the logic to subscribe to the db and send failed or enqueued pixels
