+++ app/src/main/java/com/duckduckgo/app/global/DuckDuckGoApplication.kt
@@ -47,6 +47,7 @@ import com.duckduckgo.app.statistics.AtbInitializer
 import com.duckduckgo.app.statistics.VariantManager
 import com.duckduckgo.app.statistics.api.OfflinePixelScheduler
 import com.duckduckgo.app.statistics.api.OfflinePixelSender
+import com.duckduckgo.app.statistics.api.PixelSender
 import com.duckduckgo.app.statistics.api.StatisticsUpdater
 import com.duckduckgo.app.statistics.pixels.Pixel
 import com.duckduckgo.app.statistics.pixels.Pixel.PixelName.APP_LAUNCH
@@ -143,6 +144,9 @@ open class DuckDuckGoApplication : HasAndroidInjector, Application(), LifecycleO
     lateinit var offlinePixelSender: OfflinePixelSender
 
     @Inject
+    lateinit var pixelSender: PixelSender
+
+    @Inject
     lateinit var alertingUncaughtExceptionHandler: AlertingUncaughtExceptionHandler
 
     @Inject
@@ -183,10 +187,7 @@ open class DuckDuckGoApplication : HasAndroidInjector, Application(), LifecycleO
             it.addObserver(appEnjoymentLifecycleObserver)
             it.addObserver(dataClearerForegroundAppRestartPixel)
             it.addObserver(userStageStore)
-            val lifecycleObserver = pixel as? LifecycleObserver
-            if (lifecycleObserver != null) {
-                it.addObserver(lifecycleObserver)
-            }
+            it.addObserver(pixelSender)
         }
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N_MR1) {
